remove sq duplication
