+++ SysML2.NET.Serializer.TextualNotation/Writers/FlowTextualNotationBuilder.cs
@@ -22,8 +22,6 @@ namespace SysML2.NET.Serializer.TextualNotation.Writers
 {
     using System.Text;
 
-    using SysML2.NET.Core.POCO.Core.Features;
-    using SysML2.NET.Core.POCO.Core.Types;
     using SysML2.NET.Core.POCO.Kernel.Interactions;
 
     /// <summary>
@@ -45,70 +43,13 @@ namespace SysML2.NET.Serializer.TextualNotation.Writers
         ///   | ( isSufficient ?= 'all' )?
         ///     ownedRelationship += FlowEndMember 'to' ownedRelationship += FlowEndMember
         ///
-        /// Auto-gen emits FeaturePrefix + 'flow ' before and TypeBody after this method.
+        /// Auto-gen emits FeaturePrefix + 'flow ' before and TypeBody after this method. Body delegates
+        /// to the shared <see cref="SharedTextualNotationBuilder.BuildFlowDeclarationHandCoded"/> helper —
+        /// the rule shares its body with <c>SuccessionFlow</c>'s <c>FlowDeclaration</c>.
         /// </remarks>
         private static void BuildFlowDeclarationHandCoded(IFlow poco, TextualNotationWriterContext writerContext, StringBuilder stringBuilder)
         {
-            var ownedRelationshipCursor = writerContext.CursorCache.GetOrCreateCursor(poco.Id, "ownedRelationship", poco.OwnedRelationship);
-
-            var hasDeclaration = !string.IsNullOrWhiteSpace(poco.DeclaredShortName)
-                                 || !string.IsNullOrWhiteSpace(poco.DeclaredName);
-
-            if (hasDeclaration || ownedRelationshipCursor.Current is not IEndFeatureMembership)
-            {
-                // Alt 1: FeatureDeclaration ValuePart? ('of' PayloadFeatureMember)? ('from' FlowEndMember 'to' FlowEndMember)?
-                FeatureTextualNotationBuilder.BuildFeatureDeclaration(poco, writerContext, stringBuilder);
-                FeatureTextualNotationBuilder.BuildValuePart(poco, writerContext, stringBuilder);
-
-                // 'of' PayloadFeatureMember? — IFeatureMembership but NOT IEndFeatureMembership
-                if (ownedRelationshipCursor.Current is IFeatureMembership payloadMember
-                    && ownedRelationshipCursor.Current is not IEndFeatureMembership)
-                {
-                    stringBuilder.Append("of ");
-                    FeatureMembershipTextualNotationBuilder.BuildPayloadFeatureMember(payloadMember, writerContext, stringBuilder);
-                    ownedRelationshipCursor.Move();
-                }
-
-                // 'from' FlowEndMember 'to' FlowEndMember?
-                if (ownedRelationshipCursor.Current is IEndFeatureMembership firstFlowEnd)
-                {
-                    stringBuilder.Append("from ");
-                    EndFeatureMembershipTextualNotationBuilder.BuildFlowEndMember(firstFlowEnd, writerContext, stringBuilder);
-                    ownedRelationshipCursor.Move();
-
-                    stringBuilder.Append("to ");
-
-                    if (ownedRelationshipCursor.Current is IEndFeatureMembership secondFlowEnd)
-                    {
-                        EndFeatureMembershipTextualNotationBuilder.BuildFlowEndMember(secondFlowEnd, writerContext, stringBuilder);
-                    }
-
-                    ownedRelationshipCursor.Move();
-                }
-            }
-            else
-            {
-                // Alt 2: (isSufficient?='all')? FlowEndMember 'to' FlowEndMember
-                if (poco.IsSufficient)
-                {
-                    stringBuilder.Append("all ");
-                }
-
-                if (ownedRelationshipCursor.Current is IEndFeatureMembership firstFlowEnd)
-                {
-                    EndFeatureMembershipTextualNotationBuilder.BuildFlowEndMember(firstFlowEnd, writerContext, stringBuilder);
-                    ownedRelationshipCursor.Move();
-
-                    stringBuilder.Append("to ");
-
-                    if (ownedRelationshipCursor.Current is IEndFeatureMembership secondFlowEnd)
-                    {
-                        EndFeatureMembershipTextualNotationBuilder.BuildFlowEndMember(secondFlowEnd, writerContext, stringBuilder);
-                    }
-
-                    ownedRelationshipCursor.Move();
-                }
-            }
+            SharedTextualNotationBuilder.BuildFlowDeclarationHandCoded(poco, writerContext, stringBuilder);
         }
     }
 }
